criando custom repo para sub comentario
